requirements.txt
python_telegram_bot~=22.3
fastapi~=0.115
uvicorn[standard]~=0.30
python-dotenv~=1.0
a2wsgi~=1.10.10
